requirements.txt
https://download.pytorch.org/whl/torch/torch-1.13.1%2Bcpu-cp39-cp39m-linux_x86_64.whl
https://download.pytorch.org/whl/torch/torchvision-0.14.1%2Bcpu-cp39-cp39m-linux_x86_64.whl
annotated-types==0.6.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blis==0.7.11
Bottleneck==1.3.8
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.11
django-allauth==0.61.1
django-braces==1.14.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.15.1
fastai==1.0.60
fastprogress==0.2.2
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
gunicorn==19.7.1
httptools==0.1.1
idna==2.8
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
multidict==4.7.5
murmurhash==1.0.10
networkx==3.2.1
numexpr==2.10.0
numpy==1.26.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2018.11.26
plac==1.1.3
platformdirs==4.2.0
preshed==3.0.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==1.7.1
pynvx==1.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.13.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
sympy==1.12
thinc==8.2.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.25.11
virtualenv==20.25.3
virtualenv-clone==0.5.7
wasabi==1.1.2
weasel==0.3.4
whitenoise==6.6.0
zipp==3.18.1
